requirements.txt
firebase-admin
djangorestframework
python-dotenv
